+++ public/3.js
@@ -1,21 +1,14 @@
 (window["webpackJsonp"] = window["webpackJsonp"] || []).push([[3],{
 
-/***/ "./node_modules/babel-loader/lib/index.js?!./node_modules/vue-loader/lib/index.js?!./resources/js/components/user/categories/desktopNavbar.vue?vue&type=script&lang=js&":
-/*!****************************************************************************************************************************************************************************************!*\
-  !*** ./node_modules/babel-loader/lib??ref--4-0!./node_modules/vue-loader/lib??vue-loader-options!./resources/js/components/user/categories/desktopNavbar.vue?vue&type=script&lang=js& ***!
-  \****************************************************************************************************************************************************************************************/
+/***/ "./node_modules/babel-loader/lib/index.js?!./node_modules/vue-loader/lib/index.js?!./resources/js/components/user/breadcrumbs/breadcrumb.vue?vue&type=script&lang=js&":
+/*!**************************************************************************************************************************************************************************************!*\
+  !*** ./node_modules/babel-loader/lib??ref--4-0!./node_modules/vue-loader/lib??vue-loader-options!./resources/js/components/user/breadcrumbs/breadcrumb.vue?vue&type=script&lang=js& ***!
+  \**************************************************************************************************************************************************************************************/
 /*! exports provided: default */
 /***/ (function(module, __webpack_exports__, __webpack_require__) {
 
 "use strict";
 __webpack_require__.r(__webpack_exports__);
-/* harmony import */ var vuex__WEBPACK_IMPORTED_MODULE_0__ = __webpack_require__(/*! vuex */ "./node_modules/vuex/dist/vuex.esm.js");
-function ownKeys(object, enumerableOnly) { var keys = Object.keys(object); if (Object.getOwnPropertySymbols) { var symbols = Object.getOwnPropertySymbols(object); if (enumerableOnly) symbols = symbols.filter(function (sym) { return Object.getOwnPropertyDescriptor(object, sym).enumerable; }); keys.push.apply(keys, symbols); } return keys; }
-
-function _objectSpread(target) { for (var i = 1; i < arguments.length; i++) { var source = arguments[i] != null ? arguments[i] : {}; if (i % 2) { ownKeys(Object(source), true).forEach(function (key) { _defineProperty(target, key, source[key]); }); } else if (Object.getOwnPropertyDescriptors) { Object.defineProperties(target, Object.getOwnPropertyDescriptors(source)); } else { ownKeys(Object(source)).forEach(function (key) { Object.defineProperty(target, key, Object.getOwnPropertyDescriptor(source, key)); }); } } return target; }
-
-function _defineProperty(obj, key, value) { if (key in obj) { Object.defineProperty(obj, key, { value: value, enumerable: true, configurable: true, writable: true }); } else { obj[key] = value; } return obj; }
-
 //
 //
 //
@@ -54,94 +47,19 @@ function _defineProperty(obj, key, value) { if (key in obj) { Object.definePrope
 //
 //
 //
-//
-//
-//
-//
-//
-//
-//
-//
-//
-//
-//
-//
-//
-//
-//
-//
-//
-//
-//
-//
-//
-//
-//
-//
-//
-//
-//
-//
-//
-//
-//
-//
-//
-//
-//
-//
-//
-//
-//
-//
-//
-//
-//
-//
-//
-//
-//
-//
-//
-//
-//
-//
-//
-//
-//
-//
-//
-//
-//
-//
-//
-//
-//
-//
-//
-//
-//
-//
-//
-//
-
 /* harmony default export */ __webpack_exports__["default"] = ({
+  props: ['show_banner'],
   mounted: function mounted() {
-    this.$store.dispatch('fetchCategories');
-  },
-  computed: _objectSpread({}, Object(vuex__WEBPACK_IMPORTED_MODULE_0__["mapState"])({
-    categories: function categories(state) {
-      return state.category.categories;
-    }
-  }))
+    console.log('mounted');
+  }
 });
 
 /***/ }),
 
-/***/ "./node_modules/vue-loader/lib/loaders/templateLoader.js?!./node_modules/vue-loader/lib/index.js?!./resources/js/components/user/categories/desktopNavbar.vue?vue&type=template&id=017869a2&":
-/*!********************************************************************************************************************************************************************************************************************************!*\
-  !*** ./node_modules/vue-loader/lib/loaders/templateLoader.js??vue-loader-options!./node_modules/vue-loader/lib??vue-loader-options!./resources/js/components/user/categories/desktopNavbar.vue?vue&type=template&id=017869a2& ***!
-  \********************************************************************************************************************************************************************************************************************************/
+/***/ "./node_modules/vue-loader/lib/loaders/templateLoader.js?!./node_modules/vue-loader/lib/index.js?!./resources/js/components/user/breadcrumbs/breadcrumb.vue?vue&type=template&id=697b512a&scoped=true&":
+/*!******************************************************************************************************************************************************************************************************************************************!*\
+  !*** ./node_modules/vue-loader/lib/loaders/templateLoader.js??vue-loader-options!./node_modules/vue-loader/lib??vue-loader-options!./resources/js/components/user/breadcrumbs/breadcrumb.vue?vue&type=template&id=697b512a&scoped=true& ***!
+  \******************************************************************************************************************************************************************************************************************************************/
 /*! exports provided: render, staticRenderFns */
 /***/ (function(module, __webpack_exports__, __webpack_require__) {
 
@@ -153,371 +71,64 @@ var render = function() {
   var _vm = this
   var _h = _vm.$createElement
   var _c = _vm._self._c || _h
-  return _c(
-    "div",
-    {
-      staticClass: "header-menu bg-indigo sticky-nav d-lg-block d-none ptb-25px"
-    },
-    [
-      _c("div", { staticClass: "container" }, [
-        _c("div", { staticClass: "row" }, [
-          _vm._m(0),
-          _vm._v(" "),
-          _c("div", { staticClass: "col-md-7  d-flex" }, [
-            _c(
-              "div",
-              { staticClass: "header-horizontal-menu align-self-center" },
-              [
-                _c(
-                  "ul",
-                  { staticClass: "menu-content" },
-                  [
-                    _vm._m(1),
-                    _vm._v(" "),
-                    _vm._l(_vm.categories, function(parent_category) {
-                      return _c(
-                        "li",
-                        {
-                          class: parent_category.sub_categories.length
-                            ? "menu-dropdown"
-                            : "menu"
-                        },
-                        [
-                          _c("a", { attrs: { href: "#" } }, [
-                            _vm._v(_vm._s(parent_category.title) + " "),
-                            parent_category.sub_categories.length
-                              ? _c("i", { staticClass: "ion-ios-arrow-down" })
-                              : _vm._e()
-                          ]),
-                          _vm._v(" "),
-                          _c(
-                            "ul",
-                            { staticClass: "mega-menu-wrap" },
-                            _vm._l(parent_category.sub_categories, function(
-                              sub_category
-                            ) {
-                              return parent_category.sub_categories.length
-                                ? _c("li", [
-                                    _c(
-                                      "ul",
-                                      [
-                                        _c(
-                                          "li",
-                                          { staticClass: "mega-menu-title" },
-                                          [
-                                            _c("a", { attrs: { href: "#" } }, [
-                                              _vm._v(_vm._s(sub_category.title))
-                                            ])
-                                          ]
-                                        ),
-                                        _vm._v(" "),
-                                        _vm._l(
-                                          sub_category.sub_categories,
-                                          function(level2_sub_category) {
-                                            return sub_category.sub_categories
-                                              ? _c("li", [
-                                                  _c(
-                                                    "a",
-                                                    {
-                                                      attrs: {
-                                                        href: "shop-list.html"
-                                                      }
-                                                    },
-                                                    [
-                                                      _vm._v(
-                                                        _vm._s(
-                                                          level2_sub_category.title
-                                                        )
-                                                      )
-                                                    ]
-                                                  )
-                                                ])
-                                              : _vm._e()
-                                          }
-                                        )
-                                      ],
-                                      2
-                                    )
-                                  ])
-                                : _vm._e()
-                            }),
-                            0
-                          )
-                        ]
-                      )
-                    }),
-                    _vm._v(" "),
-                    _vm._m(2),
-                    _vm._v(" "),
-                    _vm._m(3),
-                    _vm._v(" "),
-                    _vm._m(4),
-                    _vm._v(" "),
-                    _vm._m(5)
-                  ],
-                  2
-                )
-              ]
-            )
-          ]),
-          _vm._v(" "),
-          _vm._m(6)
-        ])
-      ])
-    ]
-  )
+  return _c("div", [
+    _vm.show_banner
+      ? _c("div", { staticClass: "name-category" }, [_vm._m(0)])
+      : _vm._e(),
+    _vm._v(" "),
+    _vm._m(1)
+  ])
 }
 var staticRenderFns = [
   function() {
     var _vm = this
     var _h = _vm.$createElement
     var _c = _vm._self._c || _h
-    return _c("div", { staticClass: "col-md-3 d-flex" }, [
-      _c("div", { staticClass: "logo align-self-center" }, [
-        _c("a", { attrs: { href: "index-2.html" } }, [
-          _c("img", {
-            staticClass: "img-responsive",
-            attrs: { src: "assets/images/logo/logo.png", alt: "logo.png" }
-          })
-        ])
-      ])
-    ])
-  },
-  function() {
-    var _vm = this
-    var _h = _vm.$createElement
-    var _c = _vm._self._c || _h
-    return _c("li", { staticClass: "active menu-dropdown" }, [
-      _c("a", { attrs: { href: "#" } }, [
-        _vm._v("Home "),
-        _c("i", { staticClass: "ion-ios-arrow-down" })
-      ]),
-      _vm._v(" "),
-      _c("ul", { staticClass: "main-sub-menu" }, [
-        _c("li", [
-          _c("a", { attrs: { href: "index-2.html" } }, [_vm._v("Home 1")])
-        ]),
-        _vm._v(" "),
-        _c("li", [
-          _c("a", { attrs: { href: "index-3.html" } }, [_vm._v("Home 2")])
-        ]),
-        _vm._v(" "),
-        _c("li", [
-          _c("a", { attrs: { href: "index-4.html" } }, [_vm._v("Home 3")])
-        ]),
-        _vm._v(" "),
-        _c("li", [
-          _c("a", { attrs: { href: "index-5.html" } }, [_vm._v("Home 4")])
-        ])
-      ])
-    ])
-  },
-  function() {
-    var _vm = this
-    var _h = _vm.$createElement
-    var _c = _vm._self._c || _h
-    return _c("li", { staticClass: "menu-dropdown" }, [
-      _c("a", { attrs: { href: "#" } }, [
-        _vm._v("Pages "),
-        _c("i", { staticClass: "ion-ios-arrow-down" })
-      ]),
-      _vm._v(" "),
-      _c("ul", { staticClass: "main-sub-menu" }, [
-        _c("li", [
-          _c("a", { attrs: { href: "about.html" } }, [_vm._v("About Page")])
-        ]),
-        _vm._v(" "),
-        _c("li", [
-          _c("a", { attrs: { href: "cart.html" } }, [_vm._v("Cart Page")])
-        ]),
-        _vm._v(" "),
-        _c("li", [
-          _c("a", { attrs: { href: "checkout.html" } }, [
-            _vm._v("Checkout Page")
-          ])
-        ]),
-        _vm._v(" "),
-        _c("li", [
-          _c("a", { attrs: { href: "compare.html" } }, [_vm._v("Compare Page")])
-        ]),
-        _vm._v(" "),
-        _c("li", [
-          _c("a", { attrs: { href: "login.html" } }, [
-            _vm._v("Login & Register Page")
-          ])
-        ]),
-        _vm._v(" "),
-        _c("li", [
-          _c("a", { attrs: { href: "my-account.html" } }, [
-            _vm._v("Account Page")
-          ])
-        ]),
-        _vm._v(" "),
-        _c("li", [
-          _c("a", { attrs: { href: "empty-cart.html" } }, [
-            _vm._v("Empty Cart Page")
-          ])
-        ]),
-        _vm._v(" "),
-        _c("li", [
-          _c("a", { attrs: { href: "404.html" } }, [_vm._v("404 Page")])
-        ]),
-        _vm._v(" "),
-        _c("li", [
-          _c("a", { attrs: { href: "privacy-policy.html" } }, [
-            _vm._v("Privacy Policy")
-          ])
-        ]),
-        _vm._v(" "),
-        _c("li", [
-          _c("a", { attrs: { href: "faq.html" } }, [_vm._v("Faq Page")])
-        ]),
-        _vm._v(" "),
-        _c("li", [
-          _c("a", { attrs: { href: "coming-soon.html" } }, [
-            _vm._v("Coming Soon Page")
-          ])
-        ]),
-        _vm._v(" "),
-        _c("li", [
-          _c("a", { attrs: { href: "wishlist.html" } }, [
-            _vm._v("Wishlist Page")
-          ])
-        ])
-      ])
-    ])
-  },
-  function() {
-    var _vm = this
-    var _h = _vm.$createElement
-    var _c = _vm._self._c || _h
-    return _c("li", { staticClass: "menu-dropdown" }, [
-      _c("a", { attrs: { href: "#" } }, [
-        _vm._v("Blog "),
-        _c("i", { staticClass: "ion-ios-arrow-down" })
-      ]),
-      _vm._v(" "),
-      _c("ul", { staticClass: "main-sub-menu" }, [
-        _c("li", { staticClass: "menu-dropdown position-static" }, [
-          _c("a", { attrs: { href: "#" } }, [
-            _vm._v("Blog Grid "),
-            _c("i", { staticClass: "ion-ios-arrow-right" })
-          ]),
-          _vm._v(" "),
-          _c("ul", { staticClass: "main-sub-menu main-sub-menu-2" }, [
-            _c("li", [
-              _c("a", { attrs: { href: "blog-grid-left-sidebar.html" } }, [
-                _vm._v("Blog Grid Left Sidebar")
-              ])
-            ]),
-            _vm._v(" "),
-            _c("li", [
-              _c("a", { attrs: { href: "blog-grid-right-sidebar.html" } }, [
-                _vm._v("Blog Grid Right Sidebar")
-              ])
-            ])
-          ])
-        ]),
-        _vm._v(" "),
-        _c("li", { staticClass: "menu-dropdown position-static" }, [
-          _c("a", { attrs: { href: "#" } }, [
-            _vm._v("Blog List "),
-            _c("i", { staticClass: "ion-ios-arrow-right" })
-          ]),
-          _vm._v(" "),
-          _c("ul", { staticClass: "main-sub-menu main-sub-menu-2" }, [
-            _c("li", [
-              _c("a", { attrs: { href: "blog-list-left-sidebar.html" } }, [
-                _vm._v("Blog List Left Sidebar")
-              ])
-            ]),
-            _vm._v(" "),
-            _c("li", [
-              _c("a", { attrs: { href: "blog-list-right-sidebar.html" } }, [
-                _vm._v("Blog List Right Sidebar")
-              ])
-            ])
-          ])
-        ]),
-        _vm._v(" "),
-        _c("li", { staticClass: "menu-dropdown position-static" }, [
-          _c("a", { attrs: { href: "#" } }, [
-            _vm._v("Blog Single "),
-            _c("i", { staticClass: "ion-ios-arrow-right" })
-          ]),
-          _vm._v(" "),
-          _c("ul", { staticClass: "main-sub-menu main-sub-menu-2" }, [
-            _c("li", [
-              _c("a", { attrs: { href: "blog-single-left-sidebar.html" } }, [
-                _vm._v("Blog Single Left Sidebar")
-              ])
-            ]),
-            _vm._v(" "),
-            _c("li", [
-              _c("a", { attrs: { href: "blog-single-right-sidebar.html" } }, [
-                _vm._v("Blog Single Right Sidbar")
-              ])
-            ])
-          ])
-        ])
-      ])
-    ])
-  },
-  function() {
-    var _vm = this
-    var _h = _vm.$createElement
-    var _c = _vm._self._c || _h
-    return _c("li", [
-      _c("a", { attrs: { href: "contact.html" } }, [_vm._v("contact Us")])
-    ])
-  },
-  function() {
-    var _vm = this
-    var _h = _vm.$createElement
-    var _c = _vm._self._c || _h
-    return _c("li", [
-      _c(
-        "a",
-        {
-          attrs: {
-            href: "javascript:void(0)",
-            "data-toggle": "modal",
-            "data-target": "#addProductModal"
-          }
-        },
-        [_vm._v("New Product")]
-      )
-    ])
-  },
-  function() {
-    var _vm = this
-    var _h = _vm.$createElement
-    var _c = _vm._self._c || _h
-    return _c("div", { staticClass: "col-md-2 align-self-center" }, [
-      _c(
-        "div",
-        { staticClass: "header-right-element d-flex justify-content-end" },
-        [
-          _c("div", { staticClass: "header-tools d-flex" }, [
-            _c("div", { staticClass: "cart-info d-flex align-self-center" }, [
-              _c(
-                "a",
-                {
-                  staticClass: "bag offcanvas-toggle",
-                  attrs: { title: "cart", href: "#offcanvas-cart" }
-                },
-                [
-                  _c("i", { staticClass: "las la-baby-carriage" }),
-                  _c("span", { staticClass: "item-total" }, [_vm._v("$50.00")]),
-                  _c("span", { staticClass: "item-count" }, [
-                    _vm._v("3 item(s)")
+    return _c("div", { staticClass: "container" }, [
+      _c("div", { staticClass: "row" }, [
+        _c("div", { staticClass: "col-md-12 " }, [
+          _c(
+            "div",
+            {
+              staticClass: "category-banner d-flex ",
+              attrs: {
+                "data-bg-image": "assets/images/banner-image/bg_breadcrumb.jpg"
+              }
+            },
+            [
+              _c("div", { staticClass: "align-self-center w-100" }, [
+                _c("div", { staticClass: "text-center" }, [
+                  _c("h2", { staticClass: "color-white " }, [
+                    _vm._v("Pet Food")
                   ])
-                ]
-              )
+                ])
+              ])
+            ]
+          )
+        ])
+      ])
+    ])
+  },
+  function() {
+    var _vm = this
+    var _h = _vm.$createElement
+    var _c = _vm._self._c || _h
+    return _c("div", { staticClass: "breadcrumb-area" }, [
+      _c("div", { staticClass: "container" }, [
+        _c("div", { staticClass: "row" }, [
+          _c("div", { staticClass: "col-md-12" }, [
+            _c("div", { staticClass: "breadcrumb-content" }, [
+              _c("ul", { staticClass: "nav" }, [
+                _c("li", [
+                  _c("a", { attrs: { href: "index-2.html" } }, [_vm._v("Home")])
+                ]),
+                _vm._v(" "),
+                _c("li", [_vm._v("Shop")])
+              ])
             ])
           ])
-        ]
-      )
+        ])
+      ])
     ])
   }
 ]
@@ -639,17 +250,17 @@ function normalizeComponent (
 
 /***/ }),
 
-/***/ "./resources/js/components/user/categories/desktopNavbar.vue":
-/*!*******************************************************************!*\
-  !*** ./resources/js/components/user/categories/desktopNavbar.vue ***!
-  \*******************************************************************/
+/***/ "./resources/js/components/user/breadcrumbs/breadcrumb.vue":
+/*!*****************************************************************!*\
+  !*** ./resources/js/components/user/breadcrumbs/breadcrumb.vue ***!
+  \*****************************************************************/
 /*! exports provided: default */
 /***/ (function(module, __webpack_exports__, __webpack_require__) {
 
 "use strict";
 __webpack_require__.r(__webpack_exports__);
-/* harmony import */ var _desktopNavbar_vue_vue_type_template_id_017869a2___WEBPACK_IMPORTED_MODULE_0__ = __webpack_require__(/*! ./desktopNavbar.vue?vue&type=template&id=017869a2& */ "./resources/js/components/user/categories/desktopNavbar.vue?vue&type=template&id=017869a2&");
-/* harmony import */ var _desktopNavbar_vue_vue_type_script_lang_js___WEBPACK_IMPORTED_MODULE_1__ = __webpack_require__(/*! ./desktopNavbar.vue?vue&type=script&lang=js& */ "./resources/js/components/user/categories/desktopNavbar.vue?vue&type=script&lang=js&");
+/* harmony import */ var _breadcrumb_vue_vue_type_template_id_697b512a_scoped_true___WEBPACK_IMPORTED_MODULE_0__ = __webpack_require__(/*! ./breadcrumb.vue?vue&type=template&id=697b512a&scoped=true& */ "./resources/js/components/user/breadcrumbs/breadcrumb.vue?vue&type=template&id=697b512a&scoped=true&");
+/* harmony import */ var _breadcrumb_vue_vue_type_script_lang_js___WEBPACK_IMPORTED_MODULE_1__ = __webpack_require__(/*! ./breadcrumb.vue?vue&type=script&lang=js& */ "./resources/js/components/user/breadcrumbs/breadcrumb.vue?vue&type=script&lang=js&");
 /* empty/unused harmony star reexport *//* harmony import */ var _node_modules_vue_loader_lib_runtime_componentNormalizer_js__WEBPACK_IMPORTED_MODULE_2__ = __webpack_require__(/*! ../../../../../node_modules/vue-loader/lib/runtime/componentNormalizer.js */ "./node_modules/vue-loader/lib/runtime/componentNormalizer.js");
 
 
@@ -659,50 +270,50 @@ __webpack_require__.r(__webpack_exports__);
 /* normalize component */
 
 var component = Object(_node_modules_vue_loader_lib_runtime_componentNormalizer_js__WEBPACK_IMPORTED_MODULE_2__["default"])(
-  _desktopNavbar_vue_vue_type_script_lang_js___WEBPACK_IMPORTED_MODULE_1__["default"],
-  _desktopNavbar_vue_vue_type_template_id_017869a2___WEBPACK_IMPORTED_MODULE_0__["render"],
-  _desktopNavbar_vue_vue_type_template_id_017869a2___WEBPACK_IMPORTED_MODULE_0__["staticRenderFns"],
+  _breadcrumb_vue_vue_type_script_lang_js___WEBPACK_IMPORTED_MODULE_1__["default"],
+  _breadcrumb_vue_vue_type_template_id_697b512a_scoped_true___WEBPACK_IMPORTED_MODULE_0__["render"],
+  _breadcrumb_vue_vue_type_template_id_697b512a_scoped_true___WEBPACK_IMPORTED_MODULE_0__["staticRenderFns"],
   false,
   null,
-  null,
+  "697b512a",
   null
   
 )
 
 /* hot reload */
 if (false) { var api; }
-component.options.__file = "resources/js/components/user/categories/desktopNavbar.vue"
+component.options.__file = "resources/js/components/user/breadcrumbs/breadcrumb.vue"
 /* harmony default export */ __webpack_exports__["default"] = (component.exports);
 
 /***/ }),
 
-/***/ "./resources/js/components/user/categories/desktopNavbar.vue?vue&type=script&lang=js&":
-/*!********************************************************************************************!*\
-  !*** ./resources/js/components/user/categories/desktopNavbar.vue?vue&type=script&lang=js& ***!
-  \********************************************************************************************/
+/***/ "./resources/js/components/user/breadcrumbs/breadcrumb.vue?vue&type=script&lang=js&":
+/*!******************************************************************************************!*\
+  !*** ./resources/js/components/user/breadcrumbs/breadcrumb.vue?vue&type=script&lang=js& ***!
+  \******************************************************************************************/
 /*! exports provided: default */
 /***/ (function(module, __webpack_exports__, __webpack_require__) {
 
 "use strict";
 __webpack_require__.r(__webpack_exports__);
-/* harmony import */ var _node_modules_babel_loader_lib_index_js_ref_4_0_node_modules_vue_loader_lib_index_js_vue_loader_options_desktopNavbar_vue_vue_type_script_lang_js___WEBPACK_IMPORTED_MODULE_0__ = __webpack_require__(/*! -!../../../../../node_modules/babel-loader/lib??ref--4-0!../../../../../node_modules/vue-loader/lib??vue-loader-options!./desktopNavbar.vue?vue&type=script&lang=js& */ "./node_modules/babel-loader/lib/index.js?!./node_modules/vue-loader/lib/index.js?!./resources/js/components/user/categories/desktopNavbar.vue?vue&type=script&lang=js&");
-/* empty/unused harmony star reexport */ /* harmony default export */ __webpack_exports__["default"] = (_node_modules_babel_loader_lib_index_js_ref_4_0_node_modules_vue_loader_lib_index_js_vue_loader_options_desktopNavbar_vue_vue_type_script_lang_js___WEBPACK_IMPORTED_MODULE_0__["default"]); 
+/* harmony import */ var _node_modules_babel_loader_lib_index_js_ref_4_0_node_modules_vue_loader_lib_index_js_vue_loader_options_breadcrumb_vue_vue_type_script_lang_js___WEBPACK_IMPORTED_MODULE_0__ = __webpack_require__(/*! -!../../../../../node_modules/babel-loader/lib??ref--4-0!../../../../../node_modules/vue-loader/lib??vue-loader-options!./breadcrumb.vue?vue&type=script&lang=js& */ "./node_modules/babel-loader/lib/index.js?!./node_modules/vue-loader/lib/index.js?!./resources/js/components/user/breadcrumbs/breadcrumb.vue?vue&type=script&lang=js&");
+/* empty/unused harmony star reexport */ /* harmony default export */ __webpack_exports__["default"] = (_node_modules_babel_loader_lib_index_js_ref_4_0_node_modules_vue_loader_lib_index_js_vue_loader_options_breadcrumb_vue_vue_type_script_lang_js___WEBPACK_IMPORTED_MODULE_0__["default"]); 
 
 /***/ }),
 
-/***/ "./resources/js/components/user/categories/desktopNavbar.vue?vue&type=template&id=017869a2&":
-/*!**************************************************************************************************!*\
-  !*** ./resources/js/components/user/categories/desktopNavbar.vue?vue&type=template&id=017869a2& ***!
-  \**************************************************************************************************/
+/***/ "./resources/js/components/user/breadcrumbs/breadcrumb.vue?vue&type=template&id=697b512a&scoped=true&":
+/*!************************************************************************************************************!*\
+  !*** ./resources/js/components/user/breadcrumbs/breadcrumb.vue?vue&type=template&id=697b512a&scoped=true& ***!
+  \************************************************************************************************************/
 /*! exports provided: render, staticRenderFns */
 /***/ (function(module, __webpack_exports__, __webpack_require__) {
 
 "use strict";
 __webpack_require__.r(__webpack_exports__);
-/* harmony import */ var _node_modules_vue_loader_lib_loaders_templateLoader_js_vue_loader_options_node_modules_vue_loader_lib_index_js_vue_loader_options_desktopNavbar_vue_vue_type_template_id_017869a2___WEBPACK_IMPORTED_MODULE_0__ = __webpack_require__(/*! -!../../../../../node_modules/vue-loader/lib/loaders/templateLoader.js??vue-loader-options!../../../../../node_modules/vue-loader/lib??vue-loader-options!./desktopNavbar.vue?vue&type=template&id=017869a2& */ "./node_modules/vue-loader/lib/loaders/templateLoader.js?!./node_modules/vue-loader/lib/index.js?!./resources/js/components/user/categories/desktopNavbar.vue?vue&type=template&id=017869a2&");
-/* harmony reexport (safe) */ __webpack_require__.d(__webpack_exports__, "render", function() { return _node_modules_vue_loader_lib_loaders_templateLoader_js_vue_loader_options_node_modules_vue_loader_lib_index_js_vue_loader_options_desktopNavbar_vue_vue_type_template_id_017869a2___WEBPACK_IMPORTED_MODULE_0__["render"]; });
+/* harmony import */ var _node_modules_vue_loader_lib_loaders_templateLoader_js_vue_loader_options_node_modules_vue_loader_lib_index_js_vue_loader_options_breadcrumb_vue_vue_type_template_id_697b512a_scoped_true___WEBPACK_IMPORTED_MODULE_0__ = __webpack_require__(/*! -!../../../../../node_modules/vue-loader/lib/loaders/templateLoader.js??vue-loader-options!../../../../../node_modules/vue-loader/lib??vue-loader-options!./breadcrumb.vue?vue&type=template&id=697b512a&scoped=true& */ "./node_modules/vue-loader/lib/loaders/templateLoader.js?!./node_modules/vue-loader/lib/index.js?!./resources/js/components/user/breadcrumbs/breadcrumb.vue?vue&type=template&id=697b512a&scoped=true&");
+/* harmony reexport (safe) */ __webpack_require__.d(__webpack_exports__, "render", function() { return _node_modules_vue_loader_lib_loaders_templateLoader_js_vue_loader_options_node_modules_vue_loader_lib_index_js_vue_loader_options_breadcrumb_vue_vue_type_template_id_697b512a_scoped_true___WEBPACK_IMPORTED_MODULE_0__["render"]; });
 
-/* harmony reexport (safe) */ __webpack_require__.d(__webpack_exports__, "staticRenderFns", function() { return _node_modules_vue_loader_lib_loaders_templateLoader_js_vue_loader_options_node_modules_vue_loader_lib_index_js_vue_loader_options_desktopNavbar_vue_vue_type_template_id_017869a2___WEBPACK_IMPORTED_MODULE_0__["staticRenderFns"]; });
+/* harmony reexport (safe) */ __webpack_require__.d(__webpack_exports__, "staticRenderFns", function() { return _node_modules_vue_loader_lib_loaders_templateLoader_js_vue_loader_options_node_modules_vue_loader_lib_index_js_vue_loader_options_breadcrumb_vue_vue_type_template_id_697b512a_scoped_true___WEBPACK_IMPORTED_MODULE_0__["staticRenderFns"]; });
