+++ public/7.js
@@ -86,8 +86,166 @@ function _defineProperty(obj, key, value) { if (key in obj) { Object.definePrope
 //
 //
 //
+//
+//
+//
+//
+//
+//
+//
+//
+//
+//
+//
+//
+//
+//
+//
+//
+//
+//
+//
+//
+//
+//
+//
+//
+//
+//
+//
+//
+//
+//
+//
+//
+//
+//
+//
+//
+//
+//
+//
+//
+//
+//
+//
+//
+//
+//
+//
+//
+//
+//
+//
+//
+//
+//
+//
+//
+//
+//
+//
+//
+//
+//
+//
+//
+//
+//
+//
+//
+//
+//
+//
+//
+//
+//
+//
+//
+//
+//
+//
+//
+//
+//
+//
+//
+//
+//
+//
+//
+//
+//
+//
+//
+//
+//
+//
+//
+//
+//
+//
+//
+//
+//
+//
+//
+//
+//
+//
+//
+//
+//
+//
+//
+//
+//
+//
+//
+//
+//
+//
+//
+//
+//
+//
+//
+//
+//
+//
+//
+//
+//
+//
+//
+//
+//
+//
+//
+//
+//
+//
+//
+//
+//
+//
+//
+//
+//
+//
+//
+//
+//
+//
+//
+//
 
 /* harmony default export */ __webpack_exports__["default"] = ({
+  data: function data() {
+    return {
+      current_step: 1
+    };
+  },
   mounted: function mounted() {},
   computed: _objectSpread({}, Object(vuex__WEBPACK_IMPORTED_MODULE_0__["mapState"])({
     categories: function categories(state) {
@@ -113,7 +271,7 @@ exports = module.exports = __webpack_require__(/*! ../../../../../node_modules/c
 
 
 // module
-exports.push([module.i, "\n.list-group-item {\n    position: relative;\n    display: block;\n    padding: 1rem 1rem;\n    text-decoration: none;\n    background-color: #fff;\n    border: 1px solid rgba(0,0,0,.125);\n    color: rgba(0,47,52,.64);\n}\n.list-group-item.active {\n    z-index: 2;\n    color: #fff;\n    background-color: rgba(0, 47, 52, 0.30);\n    border: 1px solid rgba(0, 47, 52, 0.30);\n}\n", ""]);
+exports.push([module.i, "\n.list-group-item {\n    position: relative;\n    display: block;\n    padding: 1rem 1rem;\n    text-decoration: none;\n    background-color: #fff;\n    border: 1px solid rgba(0, 0, 0, .125);\n    color: rgba(0, 47, 52, .64);\n}\n.list-group-item.active {\n    z-index: 2;\n    color: #fff;\n    background-color: rgba(0, 47, 52, 0.30);\n    border: 1px solid rgba(0, 47, 52, 0.30);\n}\n* {\n    margin: 0;\n    padding: 0\n}\nhtml {\n    height: 100%\n}\n#grad1 {\n    background-color: #9C27B0;\n    background-image: linear-gradient(120deg, #FF4081, #81D4FA)\n}\n#msform {\n    text-align: center;\n    position: relative;\n    margin-top: 20px\n}\n#msform fieldset .form-card {\n    background: white;\n    border: 0 none;\n    border-radius: 0px;\n    box-shadow: 0 2px 2px 2px rgba(0, 0, 0, 0.2);\n    padding: 20px 40px 30px 40px;\n    box-sizing: border-box;\n    width: 94%;\n    margin: 0 3% 20px 3%;\n    position: relative\n}\n#msform fieldset {\n    background: white;\n    border: 0 none;\n    border-radius: 0.5rem;\n    box-sizing: border-box;\n    width: 100%;\n    margin: 0;\n    padding-bottom: 20px;\n    position: relative\n}\n#msform fieldset:not(:first-of-type) {\n    display: none\n}\n#msform fieldset .form-card {\n    text-align: left;\n    color: #9E9E9E\n}\n#msform input,\n#msform textarea {\n    padding: 0px 8px 4px 8px;\n    border: none;\n    border-bottom: 1px solid #ccc;\n    border-radius: 0px;\n    margin-bottom: 25px;\n    margin-top: 2px;\n    width: 100%;\n    box-sizing: border-box;\n    font-family: montserrat;\n    color: #2C3E50;\n    font-size: 16px;\n    letter-spacing: 1px\n}\n#msform input:focus,\n#msform textarea:focus {\n    box-shadow: none !important;\n    border: none;\n    font-weight: bold;\n    border-bottom: 2px solid skyblue;\n    outline-width: 0\n}\n#msform .action-button {\n    width: 100px;\n    background: skyblue;\n    font-weight: bold;\n    color: white;\n    border: 0 none;\n    border-radius: 0px;\n    /*cursor: pointer;*/\n    padding: 10px 5px;\n    margin: 10px 5px\n}\n#msform .action-button:hover,\n#msform .action-button:focus {\n    box-shadow: 0 0 0 2px white, 0 0 0 3px skyblue\n}\n#msform .action-button-previous {\n    width: 100px;\n    background: #616161;\n    font-weight: bold;\n    color: white;\n    border: 0 none;\n    border-radius: 0px;\n    cursor: pointer;\n    padding: 10px 5px;\n    margin: 10px 5px\n}\n#msform .action-button-previous:hover,\n#msform .action-button-previous:focus {\n    box-shadow: 0 0 0 2px white, 0 0 0 3px #616161\n}\nselect.list-dt {\n    border: none;\n    outline: 0;\n    border-bottom: 1px solid #ccc;\n    padding: 2px 5px 3px 5px;\n    margin: 2px\n}\nselect.list-dt:focus {\n    border-bottom: 2px solid skyblue\n}\n.card {\n    z-index: 0;\n    border: none;\n    border-radius: 0.5rem;\n    position: relative\n}\n.fs-title {\n    font-size: 25px;\n    color: #2C3E50;\n    margin-bottom: 10px;\n    font-weight: bold;\n    text-align: left\n}\n#progressbar {\n    margin-bottom: 30px;\n    overflow: hidden;\n    color: lightgrey\n}\n#progressbar .active {\n    color: #000000\n}\n#progressbar li {\n    list-style-type: none;\n    font-size: 12px;\n    width: 25%;\n    float: left;\n    position: relative\n}\n#progressbar #account:before {\n    font-family: FontAwesome;\n    content: \"\\F023\"\n}\n#progressbar #personal:before {\n    font-family: FontAwesome;\n    content: \"\\F007\"\n}\n#progressbar #payment:before {\n    font-family: FontAwesome;\n    content: \"\\F09D\"\n}\n#progressbar #confirm:before {\n    font-family: FontAwesome;\n    content: \"\\F00C\"\n}\n#progressbar li:before {\n    width: 50px;\n    height: 50px;\n    line-height: 45px;\n    display: block;\n    font-size: 18px;\n    color: #ffffff;\n    background: lightgray;\n    border-radius: 50%;\n    margin: 0 auto 10px auto;\n    padding: 2px\n}\n#progressbar li:after {\n    content: '';\n    width: 100%;\n    height: 2px;\n    background: lightgray;\n    position: absolute;\n    left: 0;\n    top: 25px;\n    z-index: -1\n}\n#progressbar li.active:before,\n#progressbar li.active:after {\n    background: skyblue\n}\n.radio-group {\n    position: relative;\n    margin-bottom: 25px\n}\n.radio {\n    display: inline-block;\n    width: 204px;\n    height: 104px;\n    border-radius: 0;\n    background: lightblue;\n    box-shadow: 0 2px 2px 2px rgba(0, 0, 0, 0.2);\n    box-sizing: border-box;\n    cursor: pointer;\n    margin: 8px 2px\n}\n.radio:hover {\n    box-shadow: 2px 2px 2px 2px rgba(0, 0, 0, 0.3)\n}\n.radio.selected {\n    box-shadow: 1px 1px 2px 2px rgba(0, 0, 0, 0.1)\n}\n.fit-image {\n    width: 100%;\n    -o-object-fit: cover;\n       object-fit: cover\n}\n.disabled{\n    cursor: not-allowed;\n}\n", ""]);
 
 // exports
 
@@ -183,95 +341,217 @@ var render = function() {
         { staticClass: "modal-dialog modal-md", attrs: { role: "document" } },
         [
           _c("div", { staticClass: "modal-content" }, [
-            _vm._m(0),
-            _vm._v(" "),
             _c("div", { staticClass: "modal-body" }, [
-              _vm._m(1),
-              _vm._v(" "),
-              _c("div", { staticClass: "row" }, [
-                _c("div", { staticClass: "col-6" }, [
+              _c(
+                "div",
+                { staticClass: "container-fluid", attrs: { id: "grad1" } },
+                [
                   _c(
                     "div",
-                    {
-                      staticClass: "list-group",
-                      attrs: { id: "list-tab", role: "tablist" }
-                    },
-                    _vm._l(_vm.categories, function(category) {
-                      return _c(
-                        "a",
-                        {
-                          class:
-                            _vm.new_product.level1_category_id === category.id
-                              ? "list-group-item list-group-item-action active"
-                              : "list-group-item list-group-item-action",
-                          attrs: {
-                            id: "category-" + category.id,
-                            "data-toggle": "list",
-                            href: "#category-" + category.id + "-sub",
-                            role: "tab",
-                            "aria-controls": "category-" + category.id + "-sub"
-                          },
-                          on: {
-                            click: function($event) {
-                              _vm.new_product.level1_category_id = category.id
-                            }
-                          }
-                        },
-                        [_vm._v(_vm._s(category.title))]
-                      )
-                    }),
-                    0
-                  )
-                ]),
-                _vm._v(" "),
-                _vm.new_product.level1_category_id > 0
-                  ? _c("div", { staticClass: "col-6" }, [
+                    { staticClass: "row justify-content-center mt-0" },
+                    [
                       _c(
                         "div",
                         {
-                          staticClass: "list-group",
-                          attrs: { id: "list-tab2", role: "tablist" }
+                          staticClass:
+                            "col-11 col-sm-9 col-md-7 col-lg-6 text-center p-0 mt-3 mb-2"
                         },
-                        _vm._l(
-                          _vm.categories.filter(function(category_) {
-                            category_.id === _vm.new_product.level1_category_id
-                          })[0].sub_categories,
-                          function(sub_category) {
-                            return _c(
-                              "a",
-                              {
-                                class:
-                                  _vm.new_product.level2_category_id ===
-                                  sub_category.id
-                                    ? "list-group-item list-group-item-action active"
-                                    : "list-group-item list-group-item-action",
-                                attrs: {
-                                  id: "category-" + sub_category.id,
-                                  "data-toggle": "list",
-                                  href: "#category-" + sub_category.id + "-sub",
-                                  role: "tab",
-                                  "aria-controls":
-                                    "category-" + sub_category.id + "-sub"
-                                },
-                                on: {
-                                  click: function($event) {
-                                    _vm.new_product.level2_category_id =
-                                      sub_category.id
-                                  }
-                                }
-                              },
-                              [_vm._v(_vm._s(sub_category.title))]
-                            )
-                          }
-                        ),
-                        0
+                        [
+                          _c(
+                            "div",
+                            { staticClass: "card px-0 pt-4 pb-0 mt-3 mb-3" },
+                            [
+                              _vm._m(0),
+                              _vm._v(" "),
+                              _c("p", [
+                                _vm._v("Fill all form field to go to next step")
+                              ]),
+                              _vm._v(" "),
+                              _c("div", { staticClass: "row" }, [
+                                _c("div", { staticClass: "col-md-12 mx-0" }, [
+                                  _c("form", { attrs: { id: "msform" } }, [
+                                    _vm._m(1),
+                                    _vm._v(" "),
+                                    _c("fieldset", [
+                                      _c("div", { staticClass: "form-card" }, [
+                                        _c("h2", { staticClass: "fs-title" }, [
+                                          _vm._v("Choose Category")
+                                        ]),
+                                        _vm._v(" "),
+                                        _c("div", { staticClass: "row" }, [
+                                          _c("div", { staticClass: "col-12" }, [
+                                            _c(
+                                              "div",
+                                              {
+                                                staticClass:
+                                                  "list-group col-6 float-left",
+                                                attrs: {
+                                                  id: "list-tab",
+                                                  role: "tablist"
+                                                }
+                                              },
+                                              _vm._l(_vm.categories, function(
+                                                category
+                                              ) {
+                                                return _c(
+                                                  "a",
+                                                  {
+                                                    class:
+                                                      _vm.new_product
+                                                        .level1_category_id ===
+                                                      category.id
+                                                        ? "list-group-item list-group-item-action active"
+                                                        : "list-group-item list-group-item-action",
+                                                    attrs: {
+                                                      id:
+                                                        "category-" +
+                                                        category.id,
+                                                      "data-toggle": "list",
+                                                      href:
+                                                        "#category-" +
+                                                        category.id +
+                                                        "-sub",
+                                                      role: "tab",
+                                                      "aria-controls":
+                                                        "category-" +
+                                                        category.id +
+                                                        "-sub"
+                                                    },
+                                                    on: {
+                                                      click: function($event) {
+                                                        _vm.new_product.level1_category_id =
+                                                          category.id
+                                                        _vm.new_product.level2_category_id = 0
+                                                      }
+                                                    }
+                                                  },
+                                                  [
+                                                    _vm._v(
+                                                      _vm._s(category.title)
+                                                    )
+                                                  ]
+                                                )
+                                              }),
+                                              0
+                                            ),
+                                            _vm._v(" "),
+                                            _vm.new_product.level1_category_id >
+                                            0
+                                              ? _c(
+                                                  "div",
+                                                  {
+                                                    staticClass:
+                                                      "list-group col-6",
+                                                    attrs: {
+                                                      id: "list-tab2",
+                                                      role: "tablist"
+                                                    }
+                                                  },
+                                                  _vm._l(
+                                                    _vm.categories.filter(
+                                                      function(category_) {
+                                                        return (
+                                                          category_.id ===
+                                                          _vm.new_product
+                                                            .level1_category_id
+                                                        )
+                                                      }
+                                                    )[0].sub_categories,
+                                                    function(sub_category) {
+                                                      return _c(
+                                                        "a",
+                                                        {
+                                                          class:
+                                                            _vm.new_product
+                                                              .level2_category_id ===
+                                                            sub_category.id
+                                                              ? "list-group-item list-group-item-action active"
+                                                              : "list-group-item list-group-item-action",
+                                                          attrs: {
+                                                            id:
+                                                              "category-" +
+                                                              sub_category.id,
+                                                            "data-toggle":
+                                                              "list",
+                                                            href:
+                                                              "#category-" +
+                                                              sub_category.id +
+                                                              "-sub",
+                                                            role: "tab",
+                                                            "aria-controls":
+                                                              "category-" +
+                                                              sub_category.id +
+                                                              "-sub"
+                                                          },
+                                                          on: {
+                                                            click: function(
+                                                              $event
+                                                            ) {
+                                                              _vm.new_product.level2_category_id =
+                                                                sub_category.id
+                                                            }
+                                                          }
+                                                        },
+                                                        [
+                                                          _vm._v(
+                                                            _vm._s(
+                                                              sub_category.title
+                                                            )
+                                                          )
+                                                        ]
+                                                      )
+                                                    }
+                                                  ),
+                                                  0
+                                                )
+                                              : _vm._e()
+                                          ])
+                                        ])
+                                      ]),
+                                      _vm._v(" "),
+                                      _c("input", {
+                                        staticClass:
+                                          "next action-button disabled",
+                                        attrs: {
+                                          type: "button",
+                                          name: "next",
+                                          value: "Next Step",
+                                          disabled:
+                                            !_vm.new_product
+                                              .level1_category_id > 0 ||
+                                            !_vm.new_product
+                                              .level2_category_id > 0
+                                        },
+                                        on: {
+                                          click: function($event) {
+                                            _vm.new_product.level1_category_id >
+                                              0 &&
+                                            _vm.new_product.level2_category_id >
+                                              0
+                                              ? (_vm.current_step = 2)
+                                              : false
+                                          }
+                                        }
+                                      })
+                                    ]),
+                                    _vm._v(" "),
+                                    _vm._m(2),
+                                    _vm._v(" "),
+                                    _vm._m(3),
+                                    _vm._v(" "),
+                                    _vm._m(4)
+                                  ])
+                                ])
+                              ])
+                            ]
+                          )
+                        ]
                       )
-                    ])
-                  : _vm._e()
-              ])
-            ]),
-            _vm._v(" "),
-            _vm._m(2)
+                    ]
+                  )
+                ]
+              )
+            ])
           ])
         ]
       )
@@ -283,58 +563,229 @@ var staticRenderFns = [
     var _vm = this
     var _h = _vm.$createElement
     var _c = _vm._self._c || _h
-    return _c("div", { staticClass: "modal-header" }, [
-      _c(
-        "h5",
-        { staticClass: "modal-title", attrs: { id: "addProductModalTitle" } },
-        [_vm._v("POST YOUR AD")]
-      ),
+    return _c("h2", [_c("strong", [_vm._v("Sign Up Your User Account")])])
+  },
+  function() {
+    var _vm = this
+    var _h = _vm.$createElement
+    var _c = _vm._self._c || _h
+    return _c("ul", { attrs: { id: "progressbar" } }, [
+      _c("li", { staticClass: "active", attrs: { id: "account" } }, [
+        _c("strong", [_vm._v("Category")])
+      ]),
       _vm._v(" "),
-      _c(
-        "button",
-        {
-          staticClass: "close",
+      _c("li", { attrs: { id: "personal" } }, [
+        _c("strong", [_vm._v("Personal")])
+      ]),
+      _vm._v(" "),
+      _c("li", { attrs: { id: "payment" } }, [
+        _c("strong", [_vm._v("Payment")])
+      ]),
+      _vm._v(" "),
+      _c("li", { attrs: { id: "confirm" } }, [_c("strong", [_vm._v("Finish")])])
+    ])
+  },
+  function() {
+    var _vm = this
+    var _h = _vm.$createElement
+    var _c = _vm._self._c || _h
+    return _c("fieldset", [
+      _c("div", { staticClass: "form-card" }, [
+        _c("h2", { staticClass: "fs-title" }, [_vm._v("Personal Information")]),
+        _vm._v(" "),
+        _c("input", {
+          attrs: { type: "text", name: "fname", placeholder: "First Name" }
+        }),
+        _vm._v(" "),
+        _c("input", {
+          attrs: { type: "text", name: "lname", placeholder: "Last Name" }
+        }),
+        _vm._v(" "),
+        _c("input", {
+          attrs: { type: "text", name: "phno", placeholder: "Contact No." }
+        }),
+        _vm._v(" "),
+        _c("input", {
           attrs: {
-            type: "button",
-            "data-dismiss": "modal",
-            "aria-label": "Close"
+            type: "text",
+            name: "phno_2",
+            placeholder: "Alternate Contact No."
           }
-        },
-        [_c("span", { attrs: { "aria-hidden": "true" } }, [_vm._v("×")])]
-      )
-    ])
-  },
-  function() {
-    var _vm = this
-    var _h = _vm.$createElement
-    var _c = _vm._self._c || _h
-    return _c("div", { staticClass: "row" }, [
-      _c("div", { staticClass: "col-12 mt-1 mb-3" }, [
-        _vm._v(
-          "\n                            CHOOSE A CATEGORY\n                        "
-        )
-      ])
-    ])
-  },
-  function() {
-    var _vm = this
-    var _h = _vm.$createElement
-    var _c = _vm._self._c || _h
-    return _c("div", { staticClass: "modal-footer" }, [
-      _c(
-        "button",
-        {
-          staticClass: "btn btn-secondary",
-          attrs: { type: "button", "data-dismiss": "modal" }
-        },
-        [_vm._v("Close")]
-      ),
+        })
+      ]),
       _vm._v(" "),
-      _c(
-        "button",
-        { staticClass: "btn btn-primary", attrs: { type: "button" } },
-        [_vm._v("Save changes")]
-      )
+      _c("input", {
+        staticClass: "previous action-button-previous",
+        attrs: { type: "button", name: "previous", value: "Previous" }
+      }),
+      _vm._v(" "),
+      _c("input", {
+        staticClass: "next action-button",
+        attrs: { type: "button", name: "next", value: "Next Step" }
+      })
+    ])
+  },
+  function() {
+    var _vm = this
+    var _h = _vm.$createElement
+    var _c = _vm._self._c || _h
+    return _c("fieldset", [
+      _c("div", { staticClass: "form-card" }, [
+        _c("h2", { staticClass: "fs-title" }, [_vm._v("Payment Information")]),
+        _vm._v(" "),
+        _c("div", { staticClass: "radio-group" }, [
+          _c(
+            "div",
+            { staticClass: "radio", attrs: { "data-value": "credit" } },
+            [
+              _c("img", {
+                attrs: {
+                  src: "https://i.imgur.com/XzOzVHZ.jpg",
+                  width: "200px",
+                  height: "100px"
+                }
+              })
+            ]
+          ),
+          _vm._v(" "),
+          _c(
+            "div",
+            { staticClass: "radio", attrs: { "data-value": "paypal" } },
+            [
+              _c("img", {
+                attrs: {
+                  src: "https://i.imgur.com/jXjwZlj.jpg",
+                  width: "200px",
+                  height: "100px"
+                }
+              })
+            ]
+          ),
+          _vm._v(" "),
+          _c("br")
+        ]),
+        _vm._v(" "),
+        _c("label", { staticClass: "pay" }, [_vm._v("Card Holder Name*")]),
+        _vm._v(" "),
+        _c("input", {
+          attrs: { type: "text", name: "holdername", placeholder: "" }
+        }),
+        _vm._v(" "),
+        _c("div", { staticClass: "row" }, [
+          _c("div", { staticClass: "col-9" }, [
+            _c("label", { staticClass: "pay" }, [_vm._v("Card Number*")]),
+            _vm._v(" "),
+            _c("input", {
+              attrs: { type: "text", name: "cardno", placeholder: "" }
+            })
+          ]),
+          _vm._v(" "),
+          _c("div", { staticClass: "col-3" }, [
+            _c("label", { staticClass: "pay" }, [_vm._v("CVC*")]),
+            _vm._v(" "),
+            _c("input", {
+              attrs: { type: "password", name: "cvcpwd", placeholder: "***" }
+            })
+          ])
+        ]),
+        _vm._v(" "),
+        _c("div", { staticClass: "row" }, [
+          _c("div", { staticClass: "col-3" }, [
+            _c("label", { staticClass: "pay" }, [_vm._v("Expiry Date*")])
+          ]),
+          _vm._v(" "),
+          _c("div", { staticClass: "col-9" }, [
+            _c(
+              "select",
+              {
+                staticClass: "list-dt",
+                attrs: { id: "month", name: "expmonth" }
+              },
+              [
+                _c("option", { attrs: { selected: "" } }, [_vm._v("Month")]),
+                _vm._v(" "),
+                _c("option", [_vm._v("January")]),
+                _vm._v(" "),
+                _c("option", [_vm._v("February")]),
+                _vm._v(" "),
+                _c("option", [_vm._v("March")]),
+                _vm._v(" "),
+                _c("option", [_vm._v("April")]),
+                _vm._v(" "),
+                _c("option", [_vm._v("May")]),
+                _vm._v(" "),
+                _c("option", [_vm._v("June")]),
+                _vm._v(" "),
+                _c("option", [_vm._v("July")]),
+                _vm._v(" "),
+                _c("option", [_vm._v("August")]),
+                _vm._v(" "),
+                _c("option", [_vm._v("September")]),
+                _vm._v(" "),
+                _c("option", [_vm._v("October")]),
+                _vm._v(" "),
+                _c("option", [_vm._v("November")]),
+                _vm._v(" "),
+                _c("option", [_vm._v("December")])
+              ]
+            ),
+            _vm._v(" "),
+            _c(
+              "select",
+              {
+                staticClass: "list-dt",
+                attrs: { id: "year", name: "expyear" }
+              },
+              [_c("option", { attrs: { selected: "" } }, [_vm._v("Year")])]
+            )
+          ])
+        ])
+      ]),
+      _vm._v(" "),
+      _c("input", {
+        staticClass: "previous action-button-previous",
+        attrs: { type: "button", name: "previous", value: "Previous" }
+      }),
+      _vm._v(" "),
+      _c("input", {
+        staticClass: "next action-button",
+        attrs: { type: "button", name: "make_payment", value: "Confirm" }
+      })
+    ])
+  },
+  function() {
+    var _vm = this
+    var _h = _vm.$createElement
+    var _c = _vm._self._c || _h
+    return _c("fieldset", [
+      _c("div", { staticClass: "form-card" }, [
+        _c("h2", { staticClass: "fs-title text-center" }, [
+          _vm._v("Success !")
+        ]),
+        _vm._v(" "),
+        _c("br"),
+        _c("br"),
+        _vm._v(" "),
+        _c("div", { staticClass: "row justify-content-center" }, [
+          _c("div", { staticClass: "col-3" }, [
+            _c("img", {
+              staticClass: "fit-image",
+              attrs: {
+                src: "https://img.icons8.com/color/96/000000/ok--v2.png"
+              }
+            })
+          ])
+        ]),
+        _vm._v(" "),
+        _c("br"),
+        _c("br"),
+        _vm._v(" "),
+        _c("div", { staticClass: "row justify-content-center" }, [
+          _c("div", { staticClass: "col-7 text-center" }, [
+            _c("h5", [_vm._v("You Have Successfully Signed Up")])
+          ])
+        ])
+      ])
     ])
   }
 ]
